// src/app/core/models/expenses.model.ts
export interface Expense {
  expensesId: string;      
  expensesDate: string;     
  voucherNo: string;
  checkNo: string;
  birReceiptNo: string;
  expensesCategory: string;
  businessName: string;
  applicationDate: string;  
  tinNo: string;
  address: string;
  expensesAmount: string;   
  vatAmount: string;
  customerName: string;
}